only save appliance.log_server to state if enabled.
by default, the REST API always include logServer.Enabled false in the response body, regardless of the given roles.
so to properly detect a diff in the state, we will only save it to the state if the log_server is enabled, so we can detect if it get deactivated.