dont use default func to initiate struct
this way, we wont send new  attributes in the payload that has been added as default values.

in 6.1 networkInactivityEnabled was added as a default value. and as a boolean this would convert as "networkInactivityEnabled: false" in the default json request body.

we don't want this since this will break backwardscompability